fix the update logic in main
Previously the character that was passed as the current was he last
typed character but this is wrong e.g. if the characters is misstyped
character it uses that character to update the matrix but the matrix
should be updated with the character coming from the sentence that was
supposed to be typed.